+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.2.0] 2021-01-21
+### Added
+- Send and receive attachments ([#8](https://github.com/gausma/nodered-contrib-signal-client/issues/8)).
+- Sender and receiver numbers in the output nodes ([#10](https://github.com/gausma/nodered-contrib-signal-client/issues/10)).
+
 ## [2.1.0] 2021-01-14
 ### Added
 - Send the receiver number in the message payload.